+++ src/content.js
@@ -48,15 +48,17 @@ async function addRecapInformation(msg) {
     }
 
     const options = await getItemsFromStorage('options');
-    if ('dismiss_class_action_info' in options && options['dismiss_class_action_info']) {
+    if (
+      'login_dismiss_new_brand_info' in options &&
+      options['login_dismiss_new_brand_info']
+    ) {
       return;
     }
 
-    const variant = await getItemsFromStorage('variant');
-    const [bannerVariant, linkVariant] = variant.split('-');
-    PACER.addRecapBannerToLoginPage(
-      bannerMessages[bannerVariant], donateLinks[linkVariant]
-    );
+    let bannerMessage =
+      'Get ready for a new chapter! RECAP is getting a ' + 'fresh new look.';
+    let bannerLink = 'https://free.law/2024/07/05/new-branding-rip-flip';
+    PACER.addRecapBannerToLoginPage(bannerMessage, bannerLink);
 
     let dismiss_button = document.getElementById('dismiss_recap_info_banner');
     dismiss_button.addEventListener('click', async () => {
@@ -66,7 +68,7 @@ async function addRecapInformation(msg) {
 
     let learn_more_btn = document.getElementById('learn_more_btn');
     learn_more_btn.addEventListener('click', async () => {
-      await PACER.removeBannerFromLoginPage();
+      await PACER.removeBannerFromLoginPage(event_from_btn = true);
       return true;
     });